backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.7.1
Flask-Cors==5.0.0
Flask-Limiter==3.9.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pytest==8.3.4
